# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreatePolicyInput:
    properties:
      allowed_licenses:
        type: string
      disallowed_deps:
        type: string
      disallowed_licenses:
        type: string
      name:
        type: string
      organisation_id:
        type: integer
    required:
    - name
    - organisation_id
    type: object
...
  handlers.UpdateProductPolicyInput:
    properties:
      policy_id:
        type: integer
    type: object
  models.Component:
    properties:
      compliance_reason:
        type: string
      compliance_status:
        description: allowed, denied, unknown
        type: string
      id:
        type: integer
      license:
        type: string
      name:
        type: string
      purl:
        type: string
      sbom_id:
        type: integer
      type:
        type: string
      version:
        type: string
      vulnerabilities:
        items:
          $ref: '#/definitions/models.Vulnerability'
        type: array
    type: object
...
  models.Policy:
    properties:
      allowed_licenses:
        description: Comma-separated
        type: string
      created_at:
        type: string
      disallowed_deps:
        description: Comma-separated (partial name match)
        type: string
      disallowed_licenses:
        description: Comma-separated
        type: string
      id:
        type: integer
      name:
        type: string
      organisation_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.SBOM:
    properties:
      components:
        items:
          $ref: '#/definitions/models.Component'
        type: array
      created_at:
        type: string
      file_name:
        type: string
      format:
        description: CycloneDX, SPDX, Syft
        type: string
      id:
        type: integer
      repository_id:
        type: integer
      updated_at:
        type: string
      version:
        type: string
    type: object
  models.Vulnerability:
    properties:
      created_at:
        type: string
      cvss:
        type: string
      details:
        type: string
      id:
        type: integer
      osv_id:
        type: string
      severity:
        description: CRITICAL, HIGH, MEDIUM, LOW
        type: string
      summary:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Scapegoat SBOM and Vulnerability Management API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scapegoat API
  version: "1.0"
paths:
...
  /policies:
    get:
      description: Get a list of policies, optionally filtered by organisation
      parameters:
      - description: Organisation ID to filter by
        in: query
        name: organisation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Policy'
            type: array
      summary: List policies
      tags:
      - policies
    post:
      consumes:
      - application/json
      description: Create a new compliance policy for an organisation
      parameters:
      - description: Policy object
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePolicyInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Policy'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new policy
      tags:
      - policies
...
  /products/{id}/policy:
    put:
      consumes:
      - application/json
      description: Assign or update a compliance policy for a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Policy ID wrapper
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProductPolicyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update product policy
      tags:
      - products
...
  /sboms/{id}:
    get:
      description: Get an SBOM by ID, including its components and vulnerabilities
      parameters:
      - description: SBOM ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SBOM'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an SBOM
      tags:
      - sboms
  /sboms/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an SBOM file (CycloneDX, SPDX, or Syft) and parse its contents.
        Optionally associate it with a repository.
      parameters:
      - description: SBOM file
        in: formData
        name: file
        required: true
        type: file
      - description: Repository ID to associate the SBOM with
        in: formData
        name: repository_id
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SBOM'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload an SBOM
      tags:
      - sboms
swagger: "2.0"